fixed small bugs
1. added linear annealing function of alpha(temperature) for sac, sac_v and tac
2. updated readme
3. fixed some little bugs.

+++ gym_wrapper/wrapper_linux.py
@@ -192,7 +192,7 @@ class gym_envs(object):
         obs, reward, done, info = list(zip(*ray.get([env.step.remote(action) for env, action in zip(self.envs, actions)])))
         self.dones_index = np.where(done)[0]
         return (np.asarray(obs),
-                np.asarray(reward),
+                np.asarray(reward).astype(np.float32),
                 np.asarray(done),
                 info)
